refactor: split files-batches page bindings
